+++ js/form.js
@@ -1,13 +1,9 @@
 import {
-  createSuccessPopup,
-  createErrorMessagePopup,
-  showMessagePopup
-} from './popup.js';
-
-import {
   MAIN_PIN_COORDINATES,
   mainMarker
 } from './map.js';
+
+import { sendData } from './data.js';
 
 const MIN_PRICES = {
   bungalow: '0',
@@ -24,8 +20,7 @@ const GUESTS_BY_ROOM = {
 const MIN_TITLE_LENGTH = 30;
 const MAX_TITLE_LENGTH = 100;
 
-const mainContent = document.querySelector('main');
-const adForm = document.querySelector('.ad-form');
+export const adForm = document.querySelector('.ad-form');
 const listingTitle = document.querySelector('#title');
 const listingTypeSelect = document.querySelector('#type');
 const pricePerNightInput = document.querySelector('#price');
@@ -123,27 +118,12 @@ adForm.addEventListener('submit', (evt) => {
   evt.preventDefault();
 
   const formData = new FormData(evt.target);
-
-  fetch('https://22.javascript.pages.academy/keksobooking',
-    {
-      method: 'POST',
-      body: formData,
-    },
-  )
-    .then((response) => {
-      if (response.ok) {
-        showMessagePopup(mainContent, createSuccessPopup());
-        adForm.reset();
-        addressField.value = `${MAIN_PIN_COORDINATES.lat}, ${MAIN_PIN_COORDINATES.lng}`;
-      } else {
-        showMessagePopup(mainContent, createErrorMessagePopup());
-      }
-    });
+  sendData(formData);
 });
 
 adForm.addEventListener('reset', () => {
   mainMarker.setLatLng(MAIN_PIN_COORDINATES);
   setTimeout(() => {
     addressField.value = `${MAIN_PIN_COORDINATES.lat}, ${MAIN_PIN_COORDINATES.lng}`;
-  }, 1)
+  }, 0)
 })
