+++ js/main.js
@@ -1,6 +1,6 @@
-import './data.js';
 import './ad-list.js';
 import './popup.js';
+import './map.js';
+import './data.js';
 import './form.js';
 import './disable.js';
-import './map.js';
