+++ js/popup.js
@@ -6,9 +6,6 @@ const LISTING_TYPES = {
 }
 
 export const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
-export const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
-export const errorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
-export const fetchErrorMessageTemplate = document.querySelector('#fetch-error').content.querySelector('.fetch-error');
 
 function getWordInRightCase(integer, word) {
   integer = Math.abs(integer) % 100;
@@ -59,26 +56,4 @@ export const createCustomPopup = ({ author, offer }) => {
   populateListingImages(offer.photos, clonedCard.querySelector('.popup__photos'));
   clonedCard.querySelector('.popup__avatar').src = author.avatar;
   return clonedCard;
-}
-
-export const createFetchErrorPopup = (err) => {
-  const clonedFetchErrorMessage = fetchErrorMessageTemplate.cloneNode(true);
-  clonedFetchErrorMessage.querySelector('.fetch-error__message').textContent = err;
-  return clonedFetchErrorMessage;
-}
-
-export const createSuccessPopup = () => {
-  const clonedSuccessMessage = successMessageTemplate.cloneNode(true);
-  return clonedSuccessMessage;
-}
-
-export const createErrorMessagePopup = () => {
-  const clonedErrorMessage = errorMessageTemplate.cloneNode(true);
-  return clonedErrorMessage;
-}
-
-export const showMessagePopup = (parent, message) => {
-  parent.appendChild(message);
-  message.addEventListener('click', () => message.remove());
-  window.addEventListener('keydown', (evt) => evt.key === 'Escape' ? message.remove() : null);
 }
